+++ .changelog/17628.txt
@@ -1,0 +1,3 @@
+```release-note:improvement
+cpustats: Use config "cpu_total_compute" (if set) for all CPU statistics
+```
