+++ client/stats/cpu_test.go
@@ -7,32 +7,17 @@ import (
 	"math"
 	"os"
 	"testing"
-	"time"
 
 	"github.com/hashicorp/nomad/ci"
-	shelpers "github.com/hashicorp/nomad/helper/stats"
 	"github.com/hashicorp/nomad/helper/testlog"
 	"github.com/stretchr/testify/assert"
 )
-
-func TestCpuStatsPercent(t *testing.T) {
-	ci.Parallel(t)
-
-	cs := NewCpuStats()
-	cs.Percent(79.7)
-	time.Sleep(1 * time.Second)
-	percent := cs.Percent(80.69)
-	expectedPercent := 98.00
-	if percent < expectedPercent && percent > (expectedPercent+1.00) {
-		t.Fatalf("expected: %v, actual: %v", expectedPercent, percent)
-	}
-}
 
 func TestHostStats_CPU(t *testing.T) {
 	ci.Parallel(t)
 
 	assert := assert.New(t)
-	assert.Nil(shelpers.Init())
+	assert.Nil(Init(0))
 
 	logger := testlog.HCLogger(t)
 	cwd, err := os.Getwd()
