+++ drivers/shared/executor/executor_basic.go
@@ -13,10 +13,10 @@ import (
 	"github.com/hashicorp/nomad/plugins/drivers"
 )
 
-func NewExecutorWithIsolation(logger hclog.Logger) Executor {
+func NewExecutorWithIsolation(logger hclog.Logger, cpuTotalTicks uint64) Executor {
 	logger = logger.Named("executor")
 	logger.Error("isolation executor is not supported on this platform, using default")
-	return NewExecutor(logger)
+	return NewExecutor(logger, cpuTotalTicks)
 }
 
 func (e *UniversalExecutor) configureResourceContainer(_ int) error { return nil }
