+++ drivers/shared/executor/utils.go
@@ -13,6 +13,7 @@ import (
 	hclog "github.com/hashicorp/go-hclog"
 	plugin "github.com/hashicorp/go-plugin"
 	"github.com/hashicorp/nomad/drivers/shared/executor/proto"
+	"github.com/hashicorp/nomad/helper/stats"
 	"github.com/hashicorp/nomad/plugins/base"
 )
 
@@ -31,6 +32,7 @@ const (
 func CreateExecutor(logger hclog.Logger, driverConfig *base.ClientDriverConfig,
 	executorConfig *ExecutorConfig) (Executor, *plugin.Client, error) {
 
+	executorConfig.CpuTotalTicks = stats.CpuTotalTicks()
 	c, err := json.Marshal(executorConfig)
 	if err != nil {
 		return nil, nil, fmt.Errorf("unable to create executor config: %v", err)
@@ -41,8 +43,9 @@ func CreateExecutor(logger hclog.Logger, driverConfig *base.ClientDriverConfig,
 	}
 
 	p := &ExecutorPlugin{
-		logger:      logger,
-		fsIsolation: executorConfig.FSIsolation,
+		logger:        logger,
+		fsIsolation:   executorConfig.FSIsolation,
+		cpuTotalTicks: executorConfig.CpuTotalTicks,
 	}
 
 	config := &plugin.ClientConfig{
@@ -75,7 +78,7 @@ func ReattachToExecutor(reattachConfig *plugin.ReattachConfig, logger hclog.Logg
 	config := &plugin.ClientConfig{
 		HandshakeConfig:  base.Handshake,
 		Reattach:         reattachConfig,
-		Plugins:          GetPluginMap(logger, false),
+		Plugins:          GetPluginMap(logger, false, stats.CpuTotalTicks()),
 		AllowedProtocols: []plugin.Protocol{plugin.ProtocolGRPC},
 		Logger:           logger.Named("executor"),
 	}
